backend/requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
sqlalchemy
python-dotenv
requests
pandas
google-genai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
